new base class for payments tests and more
