+++ tests/payments_api/test_payment.py
@@ -1,33 +1,24 @@
 import unittest
 import uuid
-import os
 from dotenv import load_dotenv
-from ping.payments_api import PaymentsApi
-from tests.test_helper import TestHelper
+from tests.payments_api.base_payments_test import BasePaymentsTest
 
 
-class TestPayment(unittest.TestCase):
+class TestPayment(BasePaymentsTest):
 
-    @classmethod
-    def setUpClass(cls):
+    def setUp(self):
         load_dotenv()
-        cls.test_helper = TestHelper
-        cls.payment_order_id = os.getenv("PAYMENT_ORDER_ID")
-        cls.payments_api = PaymentsApi(os.getenv("TENANT_ID"))
-        cls.dummy_body = cls.test_helper.get_payment_body()
+        super(TestPayment, self).setUp()
 
 # Get Payments Tests
     # gets payment correctly
     def test_get_payment_200(self):
-        payment_id = os.getenv("PAYMENT_ID")
-
-        response = self.payments_api.payment.get(self.payment_order_id, payment_id)
+        response = self.payments_api.payment.get(self.payment_order_id, self.payment_id)
         self.test_helper.run_tests(self, response, 200)
 
     # gets payment with an incorrect id
     def test_get_payment_404(self):
         payment_id = ""
-
         response = self.payments_api.payment.get(self.payment_order_id, payment_id)
         self.test_helper.run_tests(self, response, 404)
 
@@ -39,17 +30,13 @@ class TestPayment(unittest.TestCase):
 
     # Initiate a payment with incorrect values inside payment object (status code 422)
     def test_initiate_payment_422(self):
-        payment_order_id = 0
         self.dummy_body["method"] = 0
-
-        response = self.payments_api.payment.initiate(self.dummy_body, payment_order_id)
+        response = self.payments_api.payment.initiate(self.dummy_body, self.payment_order_id)
         self.test_helper.run_tests(self, response, 422)
 
     # Initiate a payment on a non-existing payment order (status code 404)
     def test_initiate_payment_404(self):
-        error_payment_order_id = uuid.uuid4()
-
-        response = self.payments_api.payment.initiate(self.dummy_body, error_payment_order_id)
+        response = self.payments_api.payment.initiate(self.dummy_body, uuid.uuid4())
         self.test_helper.run_tests(self, response, 404)
